AHK V2 Part 37

+++ Sources/Framework/Utils.ahk
@@ -32,40 +32,41 @@ getControllerState(configuration := kUndefined) {
 	pid := ProcessExist("Simulator Controller.exe")
 
 	if (load && !pid && (configuration || !FileExist(kTempDirectory . "Simulator Controller.state")))
-		try {
-			if configuration {
-				fileName := temporaryFileName("Config", "ini")
+		if FileExist(kUserConfigDirectory . "Simulator Controller.install")
+			try {
+				if configuration {
+					fileName := temporaryFileName("Config", "ini")
 
-				writeMultiMap(fileName, configuration)
+					writeMultiMap(fileName, configuration)
 
-				options := (" -Configuration `"" . fileName . "`"")
+					options := (" -Configuration `"" . fileName . "`"")
+				}
+				else
+					options := ""
+
+				exePath := ("`"" . kBinariesDirectory . "Simulator Controller.exe`" -NoStartup -NoUpdate" .  options)
+
+				Run(exePath, kBinariesDirectory, , &pid)
+
+				Sleep(1000)
+
+				tries := 30
+
+				while (tries > 0) {
+					Sleep(200)
+
+					if !ProcessExist(pid)
+						break
+				}
+
+				if configuration
+					deleteFile(fileName)
 			}
-			else
-				options := ""
+			catch Any as exception {
+				logMessage(kLogCritical, translate("Cannot start Simulator Controller (") . exePath . translate(") - please rebuild the applications in the binaries folder (") . kBinariesDirectory . translate(")"))
 
-			exePath := ("`"" . kBinariesDirectory . "Simulator Controller.exe`" -NoStartup -NoUpdate" .  options)
-
-			Run(exePath, kBinariesDirectory, , &pid)
-
-			Sleep(1000)
-
-			tries := 30
-
-			while (tries > 0) {
-				Sleep(200)
-
-				if !ProcessExist(pid)
-					break
+				return newMultiMap()
 			}
-
-			if configuration
-				deleteFile(fileName)
-		}
-		catch Any as exception {
-			logMessage(kLogCritical, translate("Cannot start Simulator Controller (") . exePath . translate(") - please rebuild the applications in the binaries folder (") . kBinariesDirectory . translate(")"))
-
-			return newMultiMap()
-		}
 
 
 	return readMultiMap(kTempDirectory . "Simulator Controller.state")
